requirements.txt.txt
streamlit
groq
sqlite3
SpeechRecognition
duckduckgo-search
pypdf
python-docx
gtts